Tablero
Modificaciones a la clase tablero:
Ya no extiende de JPanel
Agregar colores al tablero

+++ tercerAvance/src/guis/PanelFondo.java
@@ -21,8 +21,8 @@ public class PanelFondo extends javax.swing.JPanel {
 
     public void paintComponent(Graphics g){
         Dimension tamano = getSize();
-       ImageIcon imagenFondo = new ImageIcon(getClass().getResource("/imagenes/piramide.jpg"));
-        g.drawImage(imagenFondo.getImage(), 0, 0, (int)this.getWidth(), this.getHeight(), null);
+        ImageIcon imagenFondo = new ImageIcon(getClass().getResource("/imagenes/piramide.jpg"));
+        g.drawImage(imagenFondo.getImage(), 0, 0, this.getWidth(), this.getHeight(), null);
         imagenFondo.setImageObserver(this);
         setOpaque(false);
